Updated `registerDirectory` to use mime types, cache headers
Previously, why `StaticFileOptions` were being correctly passed to the `registerDirectory()` function, they weren't being utilized, as a new `StaticFileOptions` object was being registered. It is correct that the `StaticFileOptions` shouldn't be shared between instances of `UseStaticFiles()`. However, the `ContentTypeProvider` (for configuring custom MIME types) and the `max-age` header (for setting the client-side cache duration) should still be set. To accomplish this, these have been moved into the local `registerDirectory()` helper function so they're reregistered on each instance of `UseStaticFiles()`.